+++ packages/_example/CHANGELOG.md
@@ -1,3 +1,50 @@
+## example [1.8.46](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.8.45...example@1.8.46) (2024-09-17)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/agent:** upgraded to 1.42.0
+* **@forestadmin/datasource-dummy:** upgraded to 1.1.26
+* **@forestadmin/datasource-mongo:** upgraded to 1.3.15
+* **@forestadmin/datasource-mongoose:** upgraded to 1.7.15
+* **@forestadmin/datasource-sequelize:** upgraded to 1.9.3
+* **@forestadmin/datasource-sql:** upgraded to 1.14.19
+* **@forestadmin/datasource-toolkit:** upgraded to 1.35.0
+
+## example [1.8.45](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.8.44...example@1.8.45) (2024-09-16)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/agent:** upgraded to 1.41.12
+* **@forestadmin/datasource-dummy:** upgraded to 1.1.25
+
+## example [1.8.44](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.8.43...example@1.8.44) (2024-09-11)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-sql:** upgraded to 1.14.18
+
+## example [1.8.43](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.8.42...example@1.8.43) (2024-09-10)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-sql:** upgraded to 1.14.17
+
 ## example [1.8.42](https://github.com/ForestAdmin/agent-nodejs/compare/example@1.8.41...example@1.8.42) (2024-09-09)