+++ packages/plugin-aws-s3/CHANGELOG.md
@@ -1,3 +1,24 @@
+## @forestadmin/plugin-aws-s3 [1.3.86](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-aws-s3@1.3.85...@forestadmin/plugin-aws-s3@1.3.86) (2024-09-17)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.48.0
+* **@forestadmin/datasource-toolkit:** upgraded to 1.35.0
+
+## @forestadmin/plugin-aws-s3 [1.3.85](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-aws-s3@1.3.84...@forestadmin/plugin-aws-s3@1.3.85) (2024-09-16)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.47.2
+
 ## @forestadmin/plugin-aws-s3 [1.3.84](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-aws-s3@1.3.83...@forestadmin/plugin-aws-s3@1.3.84) (2024-09-09)