+++ packages/plugin-export-advanced/CHANGELOG.md
@@ -1,3 +1,24 @@
+## @forestadmin/plugin-export-advanced [1.0.97](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-export-advanced@1.0.96...@forestadmin/plugin-export-advanced@1.0.97) (2024-09-17)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.48.0
+* **@forestadmin/datasource-toolkit:** upgraded to 1.35.0
+
+## @forestadmin/plugin-export-advanced [1.0.96](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-export-advanced@1.0.95...@forestadmin/plugin-export-advanced@1.0.96) (2024-09-16)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-customizer:** upgraded to 1.47.2
+
 ## @forestadmin/plugin-export-advanced [1.0.95](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/plugin-export-advanced@1.0.94...@forestadmin/plugin-export-advanced@1.0.95) (2024-09-09)